+++ lib/doggo/storybook/property_list.ex
@@ -2,6 +2,9 @@ defmodule Doggo.Storybook.PropertyList do
   @moduledoc false
   alias PhoenixStorybook.Stories.Variation
 
+  def container, do: {:div, class: "container"}
+  def layout, do: :one_column
+
   def variations(_opts) do
     [
       %Variation{
@@ -19,8 +22,12 @@ defmodule Doggo.Storybook.PropertyList do
 
   defp slots do
     [
-      ~s(<:prop label="Name">George</:prop>),
-      ~s(<:prop label="Age">42</:prop>)
+      ~s|<:prop label="Name">Apollo</:prop>|,
+      ~s|<:prop label="Age">5</:prop>|,
+      ~s|<:prop label="Breed">Golden Retriever</:prop>|,
+      ~s|<:prop label="Favorite Toy">Frisbee</:prop>|,
+      ~s|<:prop label="Owner">Katrin Dinkelschrot</:prop>|,
+      ~s|<:prop label="Health Conditions and Dietary Requirements">Apollo has a sensitive stomach and requires a special diet, including grain-free food and regular vet check-ups.</:prop>|
     ]
   end
 end